Add transfer CLI command tests
